+++ gateway/src/redis_io/stream/xadd.rs
@@ -1,4 +1,3 @@
-use crate::backend::commands::BackendCommands as BC;
 use crate::backend::commands::{
     ChooseColorPrefBackendCommand, CreateGameBackendCommand, FindPublicGameBackendCommand,
     JoinPrivateGameBackendCommand, ReqSyncBackendCommand, SessionDisconnected,
@@ -8,8 +7,7 @@ use crate::topics;
 use micro_model_bot::gateway::AttachBot;
 
 use crate::backend::commands::IntoShared;
-use crossbeam_channel::{select, Receiver};
-use log::{error, info};
+use log::error;
 use std::sync::Arc;
 
 pub trait XAddCommands {
@@ -22,35 +20,6 @@ pub trait XAddCommands {
     fn xadd_create_game(&self, command: CreateGameBackendCommand);
     fn xadd_choose_color_pref(&self, command: ChooseColorPrefBackendCommand);
     fn xadd_session_disconnected(&self, command: SessionDisconnected);
-}
-
-pub fn start(commands_out: Receiver<BC>, cmds: &dyn XAddCommands) {
-    loop {
-        select! {
-            recv(commands_out) -> backend_command_msg => match backend_command_msg {
-                Err(e) => error!("backend command xadd {:?}",e),
-                Ok(command) => {
-                    match &command {
-                        BC::ClientHeartbeat(_) => (), // no one cares
-                        _ => info!("Send command: {:?}", &command)
-                    }
-                    match command {
-                        BC::AttachBot(attach_bot) => cmds.xadd_attach_bot(attach_bot),
-                        BC::MakeMove(c) => cmds.xadd_make_move(c),
-                        BC::ClientHeartbeat(_) => (),
-                        BC::ProvideHistory(ph) => cmds.xadd_provide_history(ph),
-                        BC::ReqSync(rs) => cmds.xadd_req_sync(rs),
-                        BC::JoinPrivateGame(jpg) => cmds.xadd_join_private_game(jpg),
-                        BC::FindPublicGame(fpg) => cmds.xadd_find_public_game(fpg),
-                        BC::CreateGame(cg) => cmds.xadd_create_game(cg),
-                        BC::ChooseColorPref(cp) => cmds.xadd_choose_color_pref(cp),
-                        BC::SessionDisconnected(sd) => cmds.xadd_session_disconnected(sd),
-                        _ => error!("cannot match backend command to xadd"),
-                    }
-                }
-            }
-        }
-    }
 }
 
 pub struct RedisXAddCommands {
@@ -192,6 +161,8 @@ mod tests {
     use super::*;
     use crate::model::*;
 
+    type BC = crate::backend::commands::BackendCommands;
+
     use crossbeam_channel::{select, unbounded, Receiver, Sender};
     use std::thread;
     use uuid::Uuid;
@@ -257,7 +228,7 @@ mod tests {
         let (test_in, test_out): (Sender<TestResult>, Receiver<TestResult>) = unbounded();
         let (cmds_in, cmds_out): (Sender<BC>, Receiver<BC>) = unbounded();
 
-        thread::spawn(move || start(cmds_out, &FakeXAddCmd { st: test_in }));
+        thread::spawn(move || super::super::write::start(cmds_out, &FakeXAddCmd { st: test_in }));
 
         cmds_in
             .send(BC::AttachBot(AttachBot {
